remobe header files from utils directory
