+++ src/config/Config.hpp
@@ -17,9 +17,18 @@
 #include <string>
 #include <vector>
 
+#include "Line.hpp"
 #include "Location.hpp"
-#include "ToString.hpp"
-#include "Types.hpp"
+#include "Utils.hpp"
+
+class Config;
+
+typedef std::vector<Line>::iterator LineIter;
+typedef std::vector<Line> LineVector;
+typedef std::vector<std::string>::iterator StringIter;
+typedef std::vector<Config> ConfigVector;
+typedef std::vector<Location> LocationVector;
+typedef std::pair<int, std::string> ErrorPage;
 
 class Config {
   friend class ConfigFile;
