+++ src/config/ConfigFile.cpp
@@ -1,7 +1,5 @@
 #include "ConfigFile.hpp"
 
-#include "Types.hpp"
-
 ConfigFile::ConfigFile() {}
 
 ConfigFile::~ConfigFile() {}
