+++ src/config/ConfigParsing.hpp
@@ -8,7 +8,6 @@
 #include "Config.hpp"
 #include "Line.hpp"
 #include "Location.hpp"
-#include "Types.hpp"
 #include "Utils.hpp"
 
 class ConfigParsing {
