+++ src/core/CgiConnector.hpp
@@ -18,9 +18,9 @@
 #include <string>
 
 #include "Location.hpp"
-#include "PrintVerbose.hpp"
 #include "Request.hpp"
 #include "Status.hpp"
+#include "Utils.hpp"
 #include "colors.hpp"
 
 #define TIMEOUT 1
