+++ src/core/Webserver.hpp
@@ -14,11 +14,11 @@
 #include <string>
 
 #include "Config.hpp"
-#include "PrintVerbose.hpp"
 #include "Request.hpp"
 #include "Response.hpp"
 #include "Signals.hpp"
 #include "Socket.hpp"
+#include "Utils.hpp"
 
 #define MAX_CLIENTS 1024
 #define BUFFER_SIZE 131072
