+++ src/http/Request.cpp
@@ -1,10 +1,5 @@
 #include "Request.hpp"
 
-#include <exception>
-
-#include "PrintVerbose.hpp"
-#include "Utils.hpp"
-
 Request::Request() {
   this->requestMethodType_ = INVALID;
   this->requestBodyExists_ = false;
