+++ src/utils/GetContentDisposition.cpp
@@ -1,6 +1,5 @@
-#include "GetContentDisposition.hpp"
-
-#include <stdexcept>
+#include "Request.hpp"
+#include "Utils.hpp"
 
 std::string getContentDisposition(const Request &request,
                                   const std::string &key) {
