+++ src/utils/Last.cpp
@@ -1,3 +1,3 @@
-#include "Last.hpp"
+#include "Utils.hpp"
 
 char last(std::string string) { return *(string.end() - 1); }
