--- src/utils/Last.hpp
@@ -1,8 +1,0 @@
-#ifndef LAST_HPP_
-#define LAST_HPP_
-
-#include <string>
-
-char last(std::string string);
-
-#endif  // LAST_HPP_
