--- src/utils/Trim.hpp
@@ -1,8 +1,0 @@
-#ifndef TRIM_HPP
-#define TRIM_HPP
-
-#include <string>
-
-std::string trim(const std::string &s, const std::string &trim);
-
-#endif  // !TRIM_HPP
